+++ BlockAlertsDemo/ToAddToYourProjects/BlockBackground.m
@@ -56,7 +56,7 @@ static BlockBackground *_sharedInstance = nil;
     return self;
 }
 
-- (unsigned)retainCount
+- (NSUInteger)retainCount
 {
     return UINT_MAX;
 }
